Renaming modules and classes to be more clear. New class names follow logic of the object model.
