+++ src/context_factory.py
@@ -1,7 +1,7 @@
 
 from src.interfaces import IMetaTrader
-from src.meta_trader_adapter import MetaTraderAdapter
-from src.meta_trader_mock import MetaTraderMock
+from src.context_mt5 import ContextMT5
+from src.context_sim import ContextSimulator
 
 class MetaTraderFactory():
 
@@ -12,6 +12,6 @@ class MetaTraderFactory():
 
     def create_meta_trader(self, json_settings: dict, credentials: dict) -> IMetaTrader:
         if (self.production):
-            return MetaTraderAdapter(json_settings, credentials)
+            return ContextMT5(json_settings, credentials)
         else:
-            return MetaTraderMock(json_settings, credentials)
+            return ContextSimulator(json_settings, credentials)
