+++ tests/test_meta_trader.py
@@ -1,10 +1,10 @@
 
 import pytest
-from src.meta_trader_mock import MetaTraderMock
+from src.context_sim import ContextSimulator
 from src.interfaces import Trade
 
 def test_meta_trader():
-    meta = MetaTraderMock()
+    meta = ContextSimulator()
     t = Trade()
     assert meta.execute_trade(t) == "Traded!"
